+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Release: 15.1.0
+Lunes 26 de Enero 2015
+* **Fix:** [Generación masiva cobros background](https://github.com/LemontechSA/ttb/pull/776)
+
 ## Hotfix: 15.0.5
 Lunes 26 de Enero 2015
 * **Fix:** [No se guarda el historial de los trabajos siempre que se llama a Cobro::GuardarCobro()](https://github.com/LemontechSA/ttb/pull/828)